refactor: merge limits files into unified limits.lib.mjs
Combines claude-limits.lib.mjs and telegram-limits.lib.mjs into a single
limits.lib.mjs file as suggested in PR review.

Changes:
- Rename claude-limits.lib.mjs to limits.lib.mjs
- Add caching layer from telegram-limits.lib.mjs to limits.lib.mjs
- Remove telegram-limits.lib.mjs
- Update imports in telegram-bot.mjs and telegram-solve-queue.lib.mjs
- Update test imports

This simplifies the codebase by having one unified file for all limits:
- System metrics: CPU, RAM, Disk
- API limits: Claude, GitHub
- Caching layer with different TTLs for API (3min) and system (2min)
